+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Next Release
 
+- Updated français translations ([#656](https://github.com/ben/foundry-ironsworn/pull/656))
 - Fix: changing the _default_ character sheet now also updates move sheets
 
 ## 1.20.26